+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.3] - 2024-11-16
+### Added
+- Data attribute data-dimbox-download-file.
+
+### Changed
+- Renamed function downloadRemoteImage to downloadRemoteFile.
+
 ## [1.1.2] - 2024-09-12
 ### Fixed
 - Fix closing DimBox on esc key press with single content.